Move integration test for CnnTrainsetBuilder to another file
